Update patch set 14
Patch Set 14:

There's a variety of issues going on here. It's not "simple to just go back to the old way of doing things". There's a lot of new plumbing added for sending native handles over binder to the codec, code that makes passing pointers between processes not a problem. Putting the camera and codec in the same process isn't going to fix whatever problems you have.

Your best bet, barring access to source code for at least the encoder OR camera HAL, is redefining the MetadataBufferType enum to totally swap NativeHandle and Camera sources. The related change for doing something similar to that is on the right track, but swapping everything over is the better choice.
Then you may have to tack on ints or pointers to the VideoNativeHandleMetadata until it matches whatever your HAL and encoder are actually passing around. (You'll need to add that into the binder plumbing as mentioned).

Patch-set: 14
Reviewer: Kyle Repinski <6998@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0