Update patch set 14
Patch Set 14:

> > > Both Hamster Tian's way and AdrianDC's way work but only on
 > AOSP
 > > > and not CM. That's our issue at this point.
 > >
 > > Well this is a major difference between AOSP and CM here:
 > > https://github.com/CyanogenMod/android_frameworks_av/commit/9501a1e1ff0a173cd85c59723ba458d07e1fca96
 > >
 > > However saying your fixes "work" is relative. Different platforms
 > > will have quirks.
 > 
 > In regards to https://github.com/CyanogenMod/android_frameworks_av/commit/9501a1e1ff0a173cd85c59723ba458d07e1fca96
 > 
 > If HAL1 must use #define LEGACY_CAM_SOURCE kMetadataBufferTypeNativeHandleSource
 > 
 > then     native_handle_close(handle);
 > native_handle_delete(handle);
 > 
 > is pretty much assassinating the work around?

No. What I hit on OMAP was *without* that change, the camera HAL became responsible for closing the native handles, which was not the case pre-N. I just know it's a major behavioral difference between AOSP and CAF/CM. That code assumes the camera HAL's metadata is dup'd, but maybe with whatever you guys are trying without altering your HAL and encoder that's not necessarily the case anymore.

Patch-set: 14