Update patch set 14
Patch Set 14:

> > > Both Hamster Tian's way and AdrianDC's way work but only on
 > AOSP
 > > > and not CM. That's our issue at this point.
 > >
 > > Well this is a major difference between AOSP and CM here:
 > > https://github.com/CyanogenMod/android_frameworks_av/commit/9501a1e1ff0a173cd85c59723ba458d07e1fca96
 > >
 > > However saying your fixes "work" is relative. Different platforms
 > > will have quirks.
 > 
 > In regards to https://github.com/CyanogenMod/android_frameworks_av/commit/9501a1e1ff0a173cd85c59723ba458d07e1fca96
 > 
 > If HAL1 must use #define LEGACY_CAM_SOURCE kMetadataBufferTypeNativeHandleSource
 > 
 > then     native_handle_close(handle);
 > native_handle_delete(handle);
 > 
 > is pretty much assassinating the work around?

Maybe this will help explain it better: https://github.com/Unlegacy-Android/android_hardware_ti_omap4/commit/f96e2f19ecb24647e2dcdc8b15f5030fada5208a

Scroll down a bit to the native_handle_(close and delete) additions. That's added in the HAL's releaseRecordingFrame in case the frameworks does NOT have the commit I'm mentioning. But maybe there's a reason that commit isn't in AOSP.

Our code is almost identical for M and N. For N, just had to axe an unnecessary extra variable from our metadata struct, and change from CameraSource to NativeHandleSource so the frameworks would decide to take care of duplicating the data to the encoder. On M it just uses CameraSource but everything else is basically the same. The CAF changes look similar but are a bit harder to read IMO.

Patch-set: 14