Update patch set 14
Patch Set 14:

> > Maybe this will help explain it better: https://github.com/Unlegacy-Android/android_hardware_ti_omap4/commit/f96e2f19ecb24647e2dcdc8b15f5030fada5208a
 > 
 > Hi Kyle,
 > I remember you saying a change like this could be implemented in
 > *either* OMX or camera HAL, but the commit you linked seems to
 > touch both areas actually.
 > I'm working on Exynos5 where I don't have access to the camera HAL
 > code right now, but the OMX stack is open source.
 > Do you happen to have an example/reference commit for such cases as
 > well?
 > If not I guess I can start digging into it myself, but I was hoping
 > to avoid reinventing the wheel.

It would be a lot more involved, but hopefully not impossible. With source code for one of the two you can see what exactly your metadata struct looks like. If it's just MetadataBufferType and native_handle_t* (or buffer_handle_t) then the other can be patched by disassembling the closed-source lib and finding where the metadata is and changing some 0s to 3s. If you link me your camera HAL I'll take a look at it!

Patch-set: 14