Update patch set 14
Patch Set 14:

> @Kyle, isnt what you did basically this? http://review.cyanogenmod.org/#/c/165214/

I hit a few extra issues. The metadata struct we used had an extra int tacked on the end which did not work anymore, but fortunately it was expendable. Then we were leaking file descriptors and running out after 30 seconds, which is when I found we were responsible for closing the native handles. Seemed like crap to me, realized CM had a frameworks patch  instead which also took care of that.

But otherwise yes, it's not too bad to fix.

Patch-set: 14