SoftAVCDec: Handle interlaced streams where two fields are sent in one input
When both fields are sent in a single input buffer, only the first field was
being decoded. Fix this by not releasing input buffer when some bytes are still
left unused in the input buffer.
If both fields are sent in a separate call, decoder handles that correctly and
returns an output frame for every two fields

Bug: 32364044
Test: Manually tested by decoding clip attached in the above bug

Change-Id: Idab5acd9cbaefc5d2560a70d375f3a532d9e78eb