Update patch set 9
Patch Set 9:

It means recording works for me with front camera, but when I exit the app (Snap) I can't open it again cause the camera seems to be in a locked state. I have to reboot to get it working again. For camcorder with back camera, I can't even get a preview there, as any app crashes immediately after switching to it.

Patch-set: 9