audio: prohibit track reusing if effect chain moves
For non offload track recycling cases, new track is created with the
same session id. Which causes effect chain existing on old output moves
to new output. If old track is still recycled here, the outcome is
effect chain and track of the same session id running on separate output
threads.

Fix is to prohibit track reusing if outputs of new track and recycled
track are not the same(hints effect chain if exists already moves).

Change-Id: I49b05621fc625932b75d4b8ff2ef2370a53e015d
CRs-Fixed: 1081913