Update patch set 1
Patch Set 1:

The Gerrit UI is a pain on mobile, but all you did was to fix the ifdef, right?
Thanks, I indeed just squashed that ifdef in to the commit before pushing, so that's probably why I didn't catch it.

JFYI everyone, this patchset should work independently of SoCs, both QC and Exynos work for me.

Patch-set: 1