AudioFlinger: increase size of client heap
To successfuly set the playback rate for most tracks, we want to
use a value of 500 for the audio sink buffer size in msec. This is
necessary to meet frame count lower bounds when increasing speed.
However, this buffer size of 500 doesn't allow for successful repeat
mode playback of 192KHz, 8 channels files as one track occupies
3MB out of 4MB of the heap. Hence, increase the heap size.

CRs-Fixed: 1073831
Change-Id: Icfbfcefd0df4e88671e3f10f0031610ef502aa42