Update patch set 8
Patch Set 8:

@Steve. Broken as in r7 broke it? I'm confused here because the r7 legacy-hal fixes are merely 2 ifdef's and TARGET_HAS_LEGACY_CAMERA_HAL1 wrapped inside of ifneq ($(TARGET_USES_MEDIA_EXTENSIONS),true) in the makefile.

@All, on msm8226 (I guess valid for all qcom) during recording I logged this: "OMXNodeInstance: useBuffer: mMetadataType[0] = -1"

So by default we get "useBackup = false" This explains why recording works without this patch. I suppose this is not the case for i9100, so (some) exynos devices may need this. 

I also guess we may want to replace/remove the commit message and add a scientific explanation if and when this gets merged.

Patch-set: 8