Decouple OMX from OMXNodeInstance.
OMXNodeInstance has a member named mOwner, which is a reference to OMX.
Since Treble conversion would replace OMX with a new class (to be named Omx),
mOwner should have a type that is different from OMX. Fortunately, the
only use of mOwner in OMXNodeInstance is to call freeNode(). This CL
introduces OMXNodeOwner, an abstract class that has freeNode() as the only
function, and changes the type of mOwner to OMXNodeOwner. This way,
OMXNodeInstance can work with both OMX and Omx.

Test: adb shell am instrument -e size small -w
'android.media.cts/android.support.test.runner.AndroidJUnitRunner'

Bug: 31399200

Change-Id: Ia2d07a48aa94da8378fd36e55244a7e7817a5f42