Update patch set 6
Patch Set 6:

Oh, I see, is it because your camera HAL is a blob, which does not have that commit linked?

It also appears this doesn't have anything to do directly the native handle functionality, so it doesn't need to be linked to that flag.

On a side note, we really need to clean up our nativehandle logic. The native handle flag logic is too linked to the HAL1 flag logic.

Patch-set: 6