Update patch set 6
Patch Set 6:

I think that Arvin uploaded this change mainly because of me when I was trying to fix camcorder issues on gemini (N camera blobs). Turns out my problem was the vidc support in the kernel. I guess that Arvin doesn't need this because he's going to wait for proper 7.1 blobs.

Patch-set: 6