Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Not sure on that one, the merge removed a portion that involves a fallback
if the mime type is not matched, might be interesting
to check if the removal is ok.

If yes, this can go in,
if not, cm-13.0 mainline might need this patch reverted.

Patch-set: 1
Label: Code-Review=0