Use libxml2 as shared lib.
The dependencies to static libraries in frameworks should be removed
as many as possible to reduce a size of the system partition.
And that will improve coverage of the VNDK libraries because this
effort might find new necessary shared libs which were linked
statically before.

Size diffs of affected binaries on sailfish-userdebug build are as
follows.

libaudiopolicymanagerdefault.so: 583440 -> 267288 (-316152)
libaudiopolicymanager.so       : 16008 -> 16004 (-4)

total                          : (-316156)

Test: building succeeded, and the image was tested on sailfish.
Bug: 33056637
Change-Id: I51d98c5c2b5d94561190a1651063737fb1cb06ff