Camera service: Switch to HIDL by default, add prop to disable
On camera service startup, use the HIDL provider codepath unless the
system property "camera.disable_treble" is set to true.

The system property is needed until the HALv1 implementation is
complete and the emulator can switch over, as well as for devices not
yet tested in passthrough mode.

Whether passthrough or separate-process HAL is used depends on the
services and libraries available on each device.

Test: Camera CTS on angler, bullhead, sailfish, ryu. HFR performance
      test on sailfish. Basic operation of camera app.
Bug: 32991422

Change-Id: I3dd580cbcddd0df2e6b1e422920aa84270286e0a