Allow mediacodec _llseek() syscall.
The mediacodec process will attempt to load one or
more HIDL HALs. Code for retrieving a handle to HAL
services recently changed to include retrieving and
parsing a manifest from the device; this parsing code
causes the _llseek system call to be executed, which
was previously not allowed.

Since mediacodec already has lseek, adding _llseek
here should be safe.

Bug: 34692351
Test: media.codec no longer crashes
Change-Id: I7f0b173522a8e88395f54c7d2630270509f7799c