Enable media metrics by default
flips the bit so that media metrics are gathered on the device by
default. This is (not yet) for the uploading of any collected metrics.

This is a standalone CL that can be reverted without any entanglements
to other changes and enhancements.  But let's hope it doesn't come to
that.

Bug: 34715856
Test: build and boot -- see collected metrics (dumpsys media.metrics)