clean up a fixed size buffer
eliminate the size of a fixed length buffer for a potentially very
long record.  Code was safe in that it didn't allow an overflow, but
we really wanted the information that was being truncated.

also some cleanup on parameters for "dumpsys media.metrics":
time units for "-since" and include a -help.

Bug: 34792286 34793404
Test: invoke dumpsys with assorted parameters to see output
Change-Id: If960afd52a509f277910e62c13d4ba60f9ed7e34