Update patch set 3
Patch Set 3:

> https://review.lineageos.org/#/c/158330/ do anything for you?

First of all, this change is for cherry-picking. And so is another change that this one depends on.
And no, the change you mentioned is not enough. That is useful for prebuilt non-pie executables. Unfortunately, we are stuck with the pre-JB camera library, that requires non-pie cameraserver in order to be loaded. Previously we disabled PIE entirely with another change, but that doesn't seem to be necessary since only cameraserver has to be non-pie compiled. So, for better security, we decided to disable pie for just this one executable.

Patch-set: 3