Update patch set 1
Patch Set 1:

> It looks like some OMX function hasn't been called yet and so we
 > fail. So shouldn't we just call the relevant function inside of OMX
 > instead of doing it in stagefright?

As far as I remember when I was debugging, ResolutionUpdate() (Exynos_OMX_H264dec.c) will get called, but it will try to set the old resolution (in a 360p->1080p change it will set 360p). I may be wrong though. Perhaps it's trying to do things too early?

A solution in the openmax repo is of course preferable, but I don't know how to do it there.

Patch-set: 1