Update patch set 2
Patch Set 2:

Looks like we are trying to do the same thing....

ciwrl and I had a miscommunication of how to merge this (merging the r78 tag was better and I wanted to keep it all consistent).  Thanks for the original push of the commits and for being so fast to fix this mix-up!

Patch-set: 2
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0