Update patch set 2
Patch Set 2:

With camera we only have blobs. Here we have the source code for the OMX codec.

As the upstream frameworks code changed it is probably that the OMX code needs to be updated. This is a work around and we should work on OMX so it will work with future version and not create a work around for something we have the source for.

Christopher linked the change which probably introduced the issue. Did someone try to revert that change and see if everything works?

If it works with a revert, check which OMX functions are called in which order. Then restore the current behaviour, check  which function are called in which order and compare it with the log before.

Patch-set: 2