Update patch set 3
Patch Set 3: Code-Review+2

Interesting, so if *someone* (not me) is able to find the Qcom commit that fixed 8974 and backport that to 8084/8960, I should be able to replicate the fix in our Exynos OMX repositories.

Ethan, I'd like to merge this change in the meantime as a workaround, do you have any objections?

Patch-set: 3
Label: Code-Review=+2