Update patch set 3
Patch Set 3:

> > > Yes, the G901F is an apq8084 device and uses all caf code under
 > > > hardware/qcom
 > > [...]
 > Ah, the 8084 branches have not been updated. This can probably
 > provide some more info if 8084 is broken, but 8974 is not.

What does this mean exactly, "the 8084 branches have not been updated"?
It would be pretty unfortunate if the apq8084 would fall back.

I'm pretty new to this stuff so I'm not sure about the layout and what
has to be done to update the branch, but I'm willing to look into that
if somebody can teach me the ropes?

Thanks,
Corinna

Patch-set: 3