Update patch set 3
Patch Set 3:

I'm on the same page with you there, I hate this commit.
But we're lacking a better solution.
The Qcom guys and girls could probably find the proper fix in media-caf somewhere, but for Exynos it's a different story.

I went through all commits in 8974 media-caf that were applied in the last few months but they either don't apply to us (fixing stupid CAF bugs and loopholes) or we already have them..

Patch-set: 3