Update patch set 1
Patch Set 1:

> I don't like shimming for a file which sources are available. ;) I
 > think it is better to change this file to add SEC_OMX support.(like
 > what we did, you may notice this is cherry-picked from a merged
 > change)

Yes, a patch that was introduced at least as early as JB, when we didn't have shims. I don't really care whether or not you _like_ to shim it or not. I'm more concerned with keeping device/oem/platform-specific code out of common repos unless its necessary.

Patch-set: 1