audiohal: Do not extract value from Result unless it is Ok
The 2-parameter version of processResult used to call the 3-param
version, and for that it was extracting the value from Result.
Now HIDL framework checks that it shouldn't be done if the transaction
wasn't OK.

Change the code so the 2-param version now only extracts the result
value if the transaction has been successful.

Bug: 35758741
Change-Id: I42cb5f5220e3a952b1c304840ccb3952c1707cd0
Test: make