stagefright: stop audio encoding before releasing writer thread
Writer thread in encoder usecase gets stuck in destructor if recorder
is stopped right after being paused. This is because encoder source is
paused when recoder pauses, then omx encoder data flow stops. This
in turn causes MediaCodecSource::read() to be blocked infinitely.

The proper sequence to exit the main thread of encoding writer is to
stop the source firstly before waiting for thread termination.

Bug:2012015
Test: Video Recording.

Change-Id: Ic70bdc2ad06d3e03f6a2966c7c7cc2557a214349
CRs-Fixed: 2012015