AAudio: do not use enum types, use sized types
This change was made at the request of the Native API Council.

Note that this could cause short term ABI incompatibilities.
But it will prevent long term incompatibilities.

Note that it did not cause a problem on Pixel XL because the size of an enum
was 4 bytes. So it matched the int32_t size.

Bug: 36062325
Test: CTS test_aaudio.cpp
Change-Id: I6bbabbdce78085716a641857ec42b0da09aaf766
Signed-off-by: Phil Burk <philburk@google.com>