Update patch set 1
Patch Set 1:

> I'm also unsure if there's some other implications of this. I
 > remember hitting this at some point when a device wasn't freeing
 > tracks fast enough.

As far as I can tell, the only implication is that it allows 4 more tracks to be opened per uid.  It seems that there are lots of games in the play store that are hitting the current max value which is causing them to have no sound at all in the game. Slightly increasing the max value fixes this.

Patch-set: 1