Fix UAF in metrics summarizer code
ASAN tools found an instance of holding a reference to a variable that
had gone out of scope, resulting in a use-after-free.
This code changes the constructs around a little and ensures that the
reference does not outlive the variable's scope.

Re-scanned other code added at the same time, but did not find any
other similar scoping issues.

Bug: 37276863
Test: boot and execute; no ASAN re-test