DO NOT MERGE Bring Metrics summarizers over from master
Bring media metrics 'summarizer' construct across from master so
that we can use it in oc-dev.
This is the functionality described in b/36736083, along
with a bug fix (b/37276863).
Simple CP didn't work because of some header file movement.

the do-not-merge is because this is already in master as a couple
of separate commits and we don't want any auto-merger complaints.
(The Merged-in line documents the commit into master)

Bug: 37915543
Test: read 'dumpsys media.metrics -summary' output
Merged-in: I325c2d2b25720d384bcc75b73c97e5b3d8fa9731
Change-Id: I8d1c29c066f6289863e99379e4bd5989af5688ea