aaudio: make openStream behavior more specific
Change API specification to state that if sampleRate, channelCount
or dataFormat are specified, then the open stream will use those values.

This change does not break the original API, which stated that the actual
values might or might not match the requested values.

The documented API now matches the actual code behavior.

Bug: 38236303
Test: unit test, test_open_params.cpp
Change-Id: Ida15546f0ee2a8cc888158f564492363a700f9ea
Signed-off-by: Phil Burk <philburk@google.com>