stagefright: put battery update under a separate lock
Battery update doesn't need to lock the main lock for
mediaplayerservice. Battery update sometimes is called
from lower level componenet (eg. StagefrightRecorder),
locking the main lock here has potential to deadlock.

Puting the battery update in a separate class so that
it's clear this is not to be mixed with the rest of
the mediaplayerservice states.

bug: 38230347
Change-Id: Idf5f26f2b07ad6303775763ce283dad0679843d5