Produce log warning when a glitch occurred.
A glitch is assumed to occur when the sum of the three
latest buffer periods cause CPU time to be unavailable
for the next buffer. For example, if a buffer period
is 4 ms, requires 3 ms of CPU time, and a running sum of
3 consecutive buffer periods is examined, this sum cannot
exceed 13 = 3*4 + (4-3), where 3*4 is the expected total
time and (4-3) is the CPU time left available by one 4 ms
period. Glitch examples:
4 + 4 + 6; 5 + 4 + 5; 2 + 2 + 10

See NBLog::Reader::alertIfGlitch.
Boolean NBLog::Reader::findGlitch should be set to true
for the notification to occur.

Note: this function needs to be developed to track glitch
occurrences over time and notify when a difference in
trend occurs.

Test: dumpsys media.log
      logcat | grep NBLog

Change-Id: Ibf120afd990a71eb40863157a28c3a5a4a610c64