Update patch set 4
Patch Set 4:

It looks like the memmove change is a no-op as well, the only thing that would change is that we're using getChannels() instead of frame->header.channels, but frame->header.channels is already validated against getChannels() (aka mStreamInfo.channels) in readBuffer(), and frame->header.channels is also what's given to us in the function parameters, so it serves as an additional check to keep it this way (imo). n.b. mStreamInfo.channels is checked elsewhere in the file to be within the correct range.

Patch-set: 4